requirements.txt
python-dotenv
pymongo
openai
langchain-openai
langgraph
scipy
typing-extensions
langchain
python-telegram-bot
scikit-learn
numpy 
sqlalchemy
psycopg2
pandas 